add gpu model and move learning into src
